feat: Make API use .env, refactored into folders
