+++ src/components/ui/skeleton.tsx
@@ -1,4 +1,5 @@
 import { cn } from "@/lib/utils"
+import React from "react";
 
 function Skeleton({
   className,
